requirements.txt
langchain>=0.1.12
langchain-community>=0.0.28
langchain-openai>=0.0.8
python-dotenv==1.0.1
supabase==1.2.0
unstructured==0.12.3
python-pptx==0.6.23
python-docx==1.1.0
openpyxl==3.1.2
pypdf==4.0.1
tiktoken==0.6.0
python-magic==0.4.27
pdf2image==1.17.0
pytesseract==0.3.10
tabulate==0.9.0
pandas==2.2.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0 
fastapi==0.115.4
uvicorn==0.31.0
pydantic==2.11.4